refactor: 이벤트 로직 분리 및 관련 모듈 정리
